+++ apps/api_test/resources.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter
 
-from extensions import Route, resp_success
-from conf.define import error_schema
+from extensions import Route, error_schema, resp_success
 from apps.api_test.logics import RedisTestResolver
 
 router = APIRouter(route_class=Route, responses=error_schema)