+++ conf/settings.py
@@ -1,4 +1,3 @@
-from typing import List
 from functools import lru_cache
 
 from pydantic import BaseModel